Cast error

+++ src/r2m2_Ae.h
@@ -1,4 +1,4 @@
 // Copyright (C) 2016 Guillaume Valadon <guillaume@valadon.net>
 
-void miasm_anal(RAnalOp_r2m2*, int, const unsigned char*, int);
+void miasm_anal(RAnalOp_r2m2*, unsigned long long, const unsigned char*, int);
 char* miasm_get_reg_profile();
